float's changed to double in most of the cases.

+++ model.c
@@ -2,11 +2,11 @@
 #ifdef SOI
 	#define K(individual, Model) (Model->coagulation_factor)
 #else
-	#define K(individual, Model) (Model->coagulation_factor * (float)((individual.radio)*(individual.radio)))
+	#define K(individual, Model) (Model->coagulation_factor * (double)((individual.radio)*(individual.radio)))
 #endif
 
-#define R(individual, Model) (Model->coagulation_radio_factor*sqrtf(sqrtf((float)(individual.size))))
+#define R(individual, Model) (Model->coagulation_radio_factor*sqrt(sqrt((double)(individual.size))))
 
-#define M(individual, Model) (Model->metabolic_factor*(float)(individual.size))
+#define M(individual, Model) (Model->metabolic_factor*(double)(individual.size))
